Move search.* metadata classes into main metadata namespace
