+++ trac-api/trac-metadata/src/main/proto/trac/metadata/search.proto
@@ -15,9 +15,9 @@
  */
 
 syntax = 'proto3';
-package trac.metadata.search;
+package trac.metadata;
 
-option java_package = "com.accenture.trac.metadata.search";
+option java_package = "com.accenture.trac.metadata";
 option java_multiple_files = true;
 
 import "trac/metadata/type.proto";
