+++ trac-services/trac-svc-meta/src/main/java/com/accenture/trac/svc/meta/dal/IMetadataDal.java
@@ -19,7 +19,7 @@ package com.accenture.trac.svc.meta.dal;
 import com.accenture.trac.metadata.ObjectType;
 import com.accenture.trac.metadata.Tag;
 import com.accenture.trac.metadata.TagSelector;
-import com.accenture.trac.metadata.search.SearchParameters;
+import com.accenture.trac.metadata.SearchParameters;
 
 import java.util.List;
 import java.util.UUID;
