+++ trac-services/trac-svc-meta/src/main/java/com/accenture/trac/svc/meta/dal/jdbc/JdbcMetadataDal.java
@@ -17,7 +17,6 @@
 package com.accenture.trac.svc.meta.dal.jdbc;
 
 import com.accenture.trac.metadata.*;
-import com.accenture.trac.metadata.search.*;
 import com.accenture.trac.common.metadata.MetadataCodec;
 import com.accenture.trac.common.exception.EStartup;
 import com.accenture.trac.common.db.JdbcDialect;
