+++ trac-services/trac-svc-meta/src/main/java/com/accenture/trac/svc/meta/dal/jdbc/JdbcSearchImpl.java
@@ -17,7 +17,7 @@
 package com.accenture.trac.svc.meta.dal.jdbc;
 
 
-import com.accenture.trac.metadata.search.SearchParameters;
+import com.accenture.trac.metadata.SearchParameters;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
