+++ trac-services/trac-svc-meta/src/main/java/com/accenture/trac/svc/meta/services/MetadataSearchService.java
@@ -17,7 +17,7 @@
 package com.accenture.trac.svc.meta.services;
 
 import com.accenture.trac.metadata.Tag;
-import com.accenture.trac.metadata.search.SearchParameters;
+import com.accenture.trac.metadata.SearchParameters;
 import com.accenture.trac.svc.meta.dal.IMetadataDal;
 
 import java.util.List;
