+++ trac-services/trac-svc-meta/src/test/java/com/accenture/trac/svc/meta/dal/MetadataDalSearchTest.java
@@ -17,7 +17,6 @@
 package com.accenture.trac.svc.meta.dal;
 
 import com.accenture.trac.metadata.*;
-import com.accenture.trac.metadata.search.*;
 import com.accenture.trac.common.metadata.TypeSystem;
 import com.accenture.trac.common.metadata.MetadataCodec;
 import com.accenture.trac.svc.meta.test.IDalTestable;
